add giving

+++ components/ItemCard.js
@@ -6,8 +6,16 @@ import CatTag from "./styles/CatTag";
 import { ItemStyles } from "./styles/ItemStyles";
 import TitleStyles from "./styles/Title";
 
-export default function ItemCard({ item }) {
-  const { name, images, id, category, description } = item;
+export default function ItemCard({ item, userid }) {
+  const {
+    name,
+    images,
+    id,
+    category,
+    description,
+    status,
+    giver: { id: giverId },
+  } = item;
   return (
     <ItemStyles>
       {images?.length > 0 ? (
@@ -25,11 +33,15 @@ export default function ItemCard({ item }) {
       </TitleStyles>
       <CatTag>{category}</CatTag>
       {description && <p>{description}</p>}
-      <div className="buttonList">
-        <Link href={{ pathname: "update", query: { id: id } }}>Edit</Link>
-        <AddToWishlist id={id} />
-        <DeleteItem id={id}>Delete</DeleteItem>
-      </div>
+      {status !== "GIVEN" && (
+        <div className="buttonList">
+          {giverId === userid && (
+            <Link href={{ pathname: "update", query: { id: id } }}>Edit</Link>
+          )}
+          {giverId !== userid && <AddToWishlist id={id} />}
+          {giverId === userid && <DeleteItem id={id}>Delete</DeleteItem>}
+        </div>
+      )}
     </ItemStyles>
   );
 }
